Scoot some more duplicated code around and enhance
Applies to #609 #944

Ref(s):
* https://gbhackers.com/owasp-a10-unvalidated-redirects-forwards/
