+++ libs/helpers.js
@@ -217,7 +217,7 @@ exports.ensureIntegerOrNull = function (aEnvVar) {
 exports.port = process.env.PORT || 8080;
 exports.securePort = process.env.SECURE_PORT || 8081;
 
-exports.baseOrigin = 'https://openuserjs.org';
+exports.baseOrigin = (isPro ? 'https://openuserjs.org' : 'http://localhost:' + exports.port); // NOTE: Watchpoint
 
 // Absolute pattern and is combined for pro and dev
 exports.patternHasSameOrigin =
@@ -253,3 +253,18 @@ exports.isSameOrigin = function (aUrl) {
     URL: url
   };
 }
+
+exports.getRedirect = function (aReq) {
+  var referer = aReq.get('Referer');
+  var redirect = '/';
+
+  if (referer) {
+    referer = url.parse(referer); // NOTE: Legacy
+    if (referer.hostname === aReq.hostname) {
+      redirect = referer.path;
+    }
+  }
+
+  return redirect;
+}
+
